scripts/requirements.txt
Jinja2==3.1.3
Pillow==10.2.0
CairoSVG==2.7.1
pathlib==1.0.1
python-dateutil==2.8.2 